MODUL-165 Change component name
