+++ src/components/component-names.ts
@@ -43,7 +43,7 @@ export const PROGRESS_NAME: string = 'm-progress';
 export const RADIO_NAME: string = 'm-radio';
 export const RADIO_GROUP_NAME: string = 'm-radio-group';
 export const RADIO_STYLE_NAME: string = 'm-radio-style';
-export const ROOT_TREE_NAME: string = 'm-root-tree';
+export const ROOT_TREE_NODE: string = 'm-tree-node';
 export const SCROLL_TOP_NAME: string = 'm-scroll-top';
 export const SESSION_EXPIRED_NAME: string = 'm-session-expired';
 export const SIDEBAR_NAME: string = 'm-sidebar';
@@ -62,4 +62,5 @@ export const TEXTFIELD_NAME: string = 'm-textfield';
 export const TIMEPICKER_NAME: string = 'm-timepicker';
 export const TOOLTIP_NAME: string = 'm-tooltip';
 export const TREE_NAME: string = 'm-tree';
+export const TREE_NODE_NAME: string = 'm-tree-node';
 export const VALIDATION_MESSAGE_NAME: string = 'm-validation-message';
