--- src/components/root-tree/root-tree.html
@@ -1,4 +1,0 @@
-<div class="m-root-tree">
-    <span v-if="isTreeEmpty()">{{ emptyTreeTxt }}</span>
-    <m-tree v-else v-for="node in tree" :node="node" :externalSelectedFile="selectedFile" @selectFile="selectFile" :openFolders="openFolders" :icon="icon"></m-tree>
-</div>
