+++ src/components/tree/tree.lang.fr.json
@@ -1,5 +1,5 @@
 {
-    "m-root-tree": {
+    "m-tree": {
         "empty": "Vide"
     }
 }
