--- src/components/tree/tree.sandbox.html
@@ -1,11 +1,0 @@
-<div>
-    <h3>Test Title</h3>
-    <p class="m-u--margin-bottom">
-        Test description
-    </p>
-    <br>
-    <br>
-    <p>
-        <m-tree :tree="testArbo" externalSelectedFile="/Dossier #1/Dossier #2/6chemin-vers-le-fichier-----------------------------.js"></m-tree>
-    </p>
-</div>
