+++ src/lang/fr.ts
@@ -31,7 +31,7 @@ const FrenchPlugin: PluginObject<any> = {
                 require('../components/modal/modal.lang.fr.json'),
                 require('../components/page-not-found/page-not-found.lang.fr.json'),
                 require('../components/phone-number/phone-number.lang.fr.json'),
-                require('../components/root-tree/root-tree.lang.fr.json'),
+                require('../components/tree/tree.lang.fr.json'),
                 require('../components/scroll-top/scroll-top.lang.fr.json'),
                 require('../components/session-expired/session-expired.lang.fr.json'),
                 require('../components/sidebar/sidebar.lang.fr.json'),
