+++ src/sandbox/index.ts
@@ -42,7 +42,6 @@ import ProgressSandboxPlugin from '../components/progress/progress.sandbox';
 import RadioGroupSandboxPlugin from '../components/radio-group/radio-group.sandbox';
 import RadioStyleSandboxPlugin from '../components/radio-style/radio-style.sandbox';
 import RadioSandboxPlugin from '../components/radio/radio.sandbox';
-import RootTreeSandboxPlugin from '../components/root-tree/root-tree.sandbox';
 import ScrollTopSandboxPlugin from '../components/scroll-top/scroll-top.sandbox';
 import SessionExpiredSandboxPlugin from '../components/session-expired/session-expired.sandbox';
 import SidebarSandboxPlugin from '../components/sidebar/sidebar.sandbox';
@@ -107,7 +106,6 @@ const SandboxesPlugin: PluginObject<any> = {
         Vue.use(RadioSandboxPlugin);
         Vue.use(RadioGroupSandboxPlugin);
         Vue.use(RadioStyleSandboxPlugin);
-        Vue.use(RootTreeSandboxPlugin);
         Vue.use(ScrollTopSandboxPlugin);
         Vue.use(SessionExpiredSandboxPlugin);
         Vue.use(SidebarSandboxPlugin);
